1.调整了采样频率为100hz a参数等 2.稳定时间400-600ms

+++ boards/lpc824/user/tasks/protocol_task.c
@@ -65,7 +65,7 @@ static int16_t protocol_get_net_weight()
  if(os_msg.status == osEventMessage){
  msg =  (task_message_t *)os_msg.value.v;
  if(msg->type == RESPONSE_NET_WEIGHT){
-   net_weight = msg->net_weight;
+   net_weight = (int16_t)msg->net_weight;
    break;
   }  
  }
